Include additional Unicode block characters (#7)
The block element characters in Monaco are all sorts of screwed up. Note
in particular the left-half block is narrower than the left three-eights
block, and the full block is narrower than the 7/8 block. The characters
are used to display progress meters, which have jarringly uneven heights
unless Menlo's are included too.